# Conflicts:
#	src/soundio/sounddeviceportaudio.cpp
#	src/soundio/sounddeviceportaudio.h
#	src/sources/soundsourceffmpeg.cpp
#	src/sources/soundsourceffmpeg.h

+++ src/effects/effectchainslot.cpp
@@ -308,7 +308,7 @@ EffectSlotPointer EffectChainSlot::addEffectSlot(const QString& group) {
 
     EffectSlotPointer pSlot(pEffectSlot);
     m_slots.append(pSlot);
-    int numEffectSlots = m_pControlNumEffectSlots->get() + 1;
+    int numEffectSlots = static_cast<int>(m_pControlNumEffectSlots->get()) + 1;
     m_pControlNumEffectSlots->forceSet(numEffectSlots);
     m_pControlChainFocusedEffect->setStates(numEffectSlots);
     return pSlot;
